Gemini Api integration ;)
