+++ pubspec.yaml
@@ -34,9 +34,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  dart_openai: ^5.0.0
   google_gemini: ^0.1.2
-  image_picker: ^1.0.7
 
 dev_dependencies:
   flutter_test:
